+++ src/App.tsx
@@ -1,7 +1,7 @@
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import Header from './components/Header';
 import Error from './Pages/ErrorPage';
-import Navbar from './components/Navbar';
+import Navbar from './components/MobileNavbar';
 import { Suspense, lazy } from 'react';
 import WatchLater from './Pages/WatchLater';
 const Trending = lazy(()=> import("./Pages/Trending"))
@@ -12,6 +12,7 @@ const TopRatedM = lazy(()=> import("./Pages/TopRatedM"))
 const Upcomingmt = lazy(()=> import("./Pages/Upcomingmt"))
 import "react-toastify/dist/ReactToastify.css";
 import { ToastContainer } from 'react-toastify';
+import Loading from './components/Loading';
 
 const App = () => {
     const routes = [
@@ -46,11 +47,7 @@ const App = () => {
         },
     ];
 
-    const loader = <div className="d-flex justify-content-center spinner-loader">
-        <div className="spinner-border text-primary" role="status">
-            <span className="visually-hidden">Loading...</span>
-        </div>
-    </div> 
+    const loader = <Loading />
 
     return (
         <div>
