+++ src/components/ModalTV.tsx
@@ -1,4 +1,4 @@
-import {  img_500, unavailable } from './Config';
+import { img_500, unavailable } from './Config';
 import { useState } from 'react';
 import useGenre from '../hooks/useGenre';
 import { ValueData } from '../Pages/Movies';
@@ -7,18 +7,18 @@ import { APIKEY } from '../Services/api-client';
 import { GoArrowLeft } from 'react-icons/go';
 import { IoClose } from 'react-icons/io5';
 
-interface Props{
+interface Props {
     show: boolean;
     isOpen: boolean;
-    page: number
-    setIsOpen: (isOpen:boolean) => void;
-    poster_path: string; 
+    page: number;
+    setIsOpen: (isOpen: boolean) => void;
+    poster_path: string;
     vote_average: number;
     title: string;
     name: string;
     media_type: string;
     overview: string;
-    first_air_date: string; 
+    first_air_date: string;
     release_date: string;
     id: number;
     value: ValueData[];
@@ -28,31 +28,46 @@ interface Video {
     key: string;
     name: string;
     type: string;
-    videos: []
-  }
+    videos: [];
+}
 
-const ModalTV = ({show, isOpen, setIsOpen,poster_path, vote_average,title,name,media_type,overview,first_air_date,release_date, id, page, value}: Props) => {
-  
-    const [trailer, setTrailer] = useState<Video>( );
+const ModalTV = ({
+    show,
+    isOpen,
+    setIsOpen,
+    poster_path,
+    vote_average,
+    title,
+    name,
+    media_type,
+    overview,
+    first_air_date,
+    release_date,
+    id,
+    page,
+    value,
+}: Props) => {
+    const [trailer, setTrailer] = useState<Video>();
     const genreURL = useGenre(value);
     const fetchTrailer = async () => {
-
         try {
-          const response = await fetch(`
+            const response = await fetch(`
           https://api.themoviedb.org/3/movie/${id}/videos?api_key=${APIKEY}&language=en-US&sort_by=popularity.desc&include_adult=false&include_video=false&page=${page}&with_genres=${genreURL}&append_to_response=videos&sort_by=vote_average.desc
           `);
-          const data = await response.json();
-          console.log(data)
-          const trailer = data.results.find((video: Video) => video.type === 'Trailer') 
-          if (trailer) {
-            setTrailer(trailer);
-          } else {
-            console.log('No trailer found', alert('no trailer found'));
-          }
+            const data = await response.json();
+            console.log(data);
+            const trailer = data.results.find(
+                (video: Video) => video.type === 'Trailer'
+            );
+            if (trailer) {
+                setTrailer(trailer);
+            } else {
+                console.log('No trailer found', alert('no trailer found'));
+            }
         } catch (error) {
-          console.error('Error fetching movie trailer', error);
+            console.error('Error fetching movie trailer', error);
         }
-      };
+    };
 
     const getColorClass = (voteAverage: number) => {
         if (voteAverage >= 7.9) {
@@ -63,49 +78,93 @@ const ModalTV = ({show, isOpen, setIsOpen,poster_path, vote_average,title,name,m
             return 'red';
         }
     };
-  
 
-  return (
-    <>
-    <div className="modal-top">
-        <button className='close-btn'  onClick={()=> setIsOpen(!isOpen)}><GoArrowLeft size="25px"/></button>
-      {show ? 
-        <><div className="modal-down">
-            <div className='modal-left'>
-              <img src={poster_path ? `${img_500 + poster_path}` : unavailable} className="poster" alt={title || name} />
-            </div>
-            <div className="details">
-              <div className="">
-                <h3 className="text-decoration-underline">{title || name}</h3>
-                <h4 className='mt-4'>Overview</h4>
-                <p className='pt-2'>{overview}</p>
-
-                <div className="">
-                  <div className='fw-bold'>Type: {media_type === "tv" ? "TV" : "Movie"}</div>
-                  <div className='fw-bold'>Release date: {first_air_date || release_date}</div>
-                </div>
-                <p className='span-para'>Ratings:<span className={getColorClass(vote_average)}> {vote_average.toFixed(1)}</span></p>
-                <button className="trailer-btn" onClick={fetchTrailer}>
-                  {trailer ? <span>Loading...</span> : <span>Play Trailer</span>}
+    return (
+        <>
+            <div className='modal-top'>
+                <button
+                    className='close-btn'
+                    onClick={() => setIsOpen(!isOpen)}>
+                    <GoArrowLeft size='25px' />
                 </button>
-              </div>
-            </div>
-          </div><CastMt movie_id={id} page={page} /></>
-      : null}
-      </div>
-      {trailer ? (
-        <><div className="modal-trailer ">
-          <button className='close-btn-trailer' onClick={() => setTrailer(undefined)}>
-          <IoClose size="35px"/>
-          </button>
-          <iframe
-            src={`https://www.youtube.com/embed/${trailer.key}`}
-            title={trailer.name}
-            allowFullScreen />
-        </div></>
-      ) : null}
-    </>
-  )
-}
+                {show ? (
+                    <>
+                        <div className='modal-down'>
+                            <div className='modal-left'>
+                                <img
+                                    src={
+                                        poster_path
+                                            ? `${img_500 + poster_path}`
+                                            : unavailable
+                                    }
+                                    className='poster'
+                                    alt={title || name}
+                                />
+                            </div>
+                            <div className='details'>
+                                <div className=''>
+                                    <h3 className='text-decoration-underline'>
+                                        {title || name}
+                                    </h3>
+                                    <h4 className='mt-4'>Overview</h4>
+                                    <p className='pt-2'>{overview}</p>
 
-export default ModalTV
+                                    <div className=''>
+                                        <div className='fw-bold'>
+                                            Type:{' '}
+                                            {media_type === 'tv'
+                                                ? 'TV'
+                                                : 'Movie'}
+                                        </div>
+                                        <div className='fw-bold'>
+                                            Release date:{' '}
+                                            {first_air_date || release_date}
+                                        </div>
+                                    </div>
+                                    <p className='span-para'>
+                                        Ratings:
+                                        <span
+                                            className={getColorClass(
+                                                vote_average
+                                            )}>
+                                            {' '}
+                                            {vote_average.toFixed(1)}
+                                        </span>
+                                    </p>
+                                    <button
+                                        className='trailer-btn'
+                                        onClick={fetchTrailer}>
+                                        {trailer ? (
+                                            <span>Loading...</span>
+                                        ) : (
+                                            <span>Play Trailer</span>
+                                        )}
+                                    </button>
+                                </div>
+                            </div>
+                        </div>
+                        <CastMt movie_id={id} page={page} />
+                    </>
+                ) : null}
+            </div>
+            {trailer ? (
+                <>
+                    <div className='modal-trailer '>
+                        <button
+                            className='close-btn-trailer'
+                            onClick={() => setTrailer(undefined)}>
+                            <IoClose size='35px' />
+                        </button>
+                        <iframe
+                            src={`https://www.youtube.com/embed/${trailer.key}`}
+                            title={trailer.name}
+                            allowFullScreen
+                        />
+                    </div>
+                </>
+            ) : null}
+        </>
+    );
+};
+
+export default ModalTV;
