+++ src/components/Navbar.tsx
@@ -1,76 +1,51 @@
-import { useState } from 'react';
-import {  NavLink } from 'react-router-dom';
-import { useSelector } from 'react-redux';
-import { RootState } from '../store';
-import { IoClose } from 'react-icons/io5';
-import { FaBars } from "react-icons/fa6";
-const Navbar = () => {
-    const [isOpen, setIsOpen] = useState(false);
+import { useSelector } from "react-redux";
+import { NavLink } from "react-router-dom";
+import { RootState } from "../store";
+
+const Footer = () => {
     const cartProducts = useSelector((state:RootState)=> state.watchlater)
     const data = [
         {
             id: '1',
-            name: 'Home',
+            name: "Home",
             link: '/',
         },
         {
             id: '2',
-            name: 'Movies',
+            name: "Movies",
             link: '/movies',
         },
         {
             id: '3',
-            name: 'TV Series',
+            name: "TV Series",
             link: '/tv',
         },
         {
             id: '5',
-            name: 'Trending',
+            name: "Trending",
             link: '/trending',
         },
-        {id:'6', name: 'Watch Later', link: '/watchlater'}
-    ];
+        {id:'6', name: 'Watch later', link: '/watchlater'}
+    ]
+   
+  return (
+    <>
+        <nav className="footer ">
+        <span className="count-later">{cartProducts.watchlater.length}</span>
+            {data.map((Val)=> (
+                <ul key={Val.id}>
+                    <li>
+                        <NavLink to={`${Val.link}`}  className='navlink'>
+                             <button className='bg-transparent text-white footer-border'>
+                                <h5 className="pt-1 fs-6 text-nowrap footer-names">{Val.name} </h5>
+                            </button>
+                        </NavLink>
+                    </li>
+                </ul>
+            ))}
+        </nav>
+    </>
+  )
+}
 
-    const handleClick = () => {
-        setIsOpen(!isOpen);
-    };
-
-    return (
-        <>
-            <div className='position-relative'>
-                <span className='bbb' onClick={handleClick}>
-                    <FaBars size="35px" className='bbb-icon'/>
-                </span>
-                <div
-                    className={isOpen ? 'dont-show' : 'show-nav'}
-                    id='navigation'>
-                    <span className='ccc' onClick={handleClick}>
-                        <IoClose size="35px" className='ccc-icon'/>
-                    </span>
-                    <nav className='navi'>
-                    <span className="count-later-nav">{cartProducts.watchlater.length}</span>
-                        <ul className='d-flex flex-column align-items-start gap-5 mt-5'>
-                            {data.map((Val) => (
-                            <li key={Val.id}>   
-                                <NavLink to={`${Val.link}`}
-                                    onClick={() => {
-                                        setIsOpen(false);
-                                    }}
-                                    className='text-decoration-none'>
-                                        <div className='d-flex flex-row gap-2 justify-content-center align-items-center text-white mx-1'>
-                                             <h5 className='pt-1 fs-6 text-nowrap'>
-                                                {Val.name} 
-                                            </h5>
-                                        </div>
-                                </NavLink>
-                            </li>
-                            ))}
-                        </ul>
-                    </nav>
-                </div>
-            </div>
-        </>
-    );
-};
-
-export default Navbar;
+export default Footer;
