+++ src/index.css
@@ -29,16 +29,14 @@ video {
 
 .modal-top::-webkit-scrollbar,
 body::-webkit-scrollbar,
-.modal-top .modal-details::-webkit-scrollbar,
-.watch-padding::-webkit-scrollbar {
+.modal-top .modal-details::-webkit-scrollbar, .watch-padding::-webkit-scrollbar {
     width: 5px;
     height: 20px;
 }
 
 .modal-top::-webkit-scrollbar-thumb,
 body::-webkit-scrollbar-thumb,
-.modal-top .modal-details::-webkit-scrollbar,
-.watch-padding::-webkit-scrollbar-thumb {
+.modal-top .modal-details::-webkit-scrollbar, .watch-padding::-webkit-scrollbar-thumb{
     background-color: white;
     border-radius: 50px;
     height: 100px;
@@ -47,14 +45,7 @@ body::-webkit-scrollbar-thumb,
 .modal-top::-webkit-scrollbar-track,
 body::-webkit-scrollbar-track,
 .modal-top .modal-details::-webkit-scrollbar {
-    background-image: radial-gradient(
-        circle,
-        #e400ff,
-        #ba00a5,
-        #82005f,
-        #440e2b,
-        #000000
-    );
+    background-image: radial-gradient(circle, #e400ff, #ba00a5, #82005f, #440e2b, #000000);
 }
 
 a {
@@ -108,7 +99,7 @@ li {
 
 .position {
     position: fixed;
-    z-index: 100;
+    z-index: 98;
     width: 100%;
     top: 0;
 }
@@ -154,15 +145,16 @@ li {
     position: relative;
 }
 
-.footer .count-later {
+.footer .count-later{
     position: absolute;
     right: 85px;
     top: 13px;
 }
 
-.others-navlink {
+.others-navlink{
     text-decoration-color: transparent;
 }
+
 
 .navlink {
     text-decoration: none;
@@ -214,8 +206,7 @@ li {
     overflow: hidden;
     border-radius: 10px;
     border: 1px solid rgba(255, 255, 255, 0.18);
-    transition: transform 0.3s ease-out, backgroud-image 0.3s ease-out,
-        opacity 0.3s ease-in-out;
+    transition: transform 0.3s ease-out, backgroud-image 0.3s ease-out, opacity 0.3s ease-in-out;
     background-color: transparent;
     background: rgba(255, 255, 255, 0.12);
     border-radius: 16px;
@@ -226,7 +217,7 @@ li {
     opacity: 0.7;
 }
 
-.cards:hover {
+.cards:hover{
     opacity: 1;
 }
 
@@ -236,6 +227,7 @@ li {
     width: 300px;
     height: 520px;
 }
+
 
 .cards span {
     position: absolute;
@@ -252,7 +244,7 @@ li {
     font-weight: 700;
 }
 
-.watch-padding {
+.watch-padding{
     padding-top: 180px;
     padding-bottom: 50px;
     height: 100dvh;
@@ -260,7 +252,7 @@ li {
     position: relative;
 }
 
-.clear-all {
+.clear-all{
     position: absolute;
     background-color: red;
     top: 16%;
@@ -274,19 +266,20 @@ li {
     align-items: center;
     gap: 5px;
 }
-.watch-remove {
-    position: absolute;
-    top: 25px;
-    right: 25px;
-    background-color: red;
-    padding: 5px;
-    border: none;
-    border-radius: 50%;
-    outline: none;
+.watch-remove{
+ position: absolute;
+ top: 25px;
+ right: 25px;
+ background-color: red;
+ padding: 5px;
+ border: none;
+ border-radius: 50%;
+ outline: none;
 }
-.watch-add {
+
+.watch-add{
     position: absolute;
-    top: 10px;
+    top: 0px;
     right: 10px;
     padding: 8px 8px;
     border-radius: 50%;
@@ -294,15 +287,15 @@ li {
     outline: none;
     background-color: transparent;
 }
-.count-later {
+.count-later{
     background-color: red;
     border-radius: 50%;
     padding: 5px 12px;
     color: white;
     font-family: Kanit, sans-serif;
-}
+   }
 
-.count-later-nav {
+.count-later-nav{
     background-color: red;
     border-radius: 50%;
     padding: 5px 12px;
@@ -313,7 +306,9 @@ li {
     color: white;
 }
 
-.display-grid-watch-later {
+
+
+.display-grid-watch-later{
     display: grid;
     grid-template-columns: repeat(auto-fit, minmax(270px, 330px));
     gap: 3rem 2rem;
@@ -434,7 +429,7 @@ a:hover {
 }
 
 .modal-top .close-btn {
-    position: absolute;
+    position: fixed;
     right: 10.2%;
     top: 15.5%;
     color: white;
@@ -464,7 +459,7 @@ a:hover {
 
 .bbb {
     position: fixed;
-    z-index: 100;
+    z-index: 98;
     top: 18px;
     left: 10px;
     border-radius: 5px;
@@ -619,7 +614,7 @@ a:hover {
 .modal-trailer {
     background-color: rgba(0, 0, 0, 0.9);
     position: fixed;
-    z-index: 102;
+    z-index: 200;
     inset: 0;
     width: 100%;
     height: 100%;
@@ -645,7 +640,7 @@ a:hover {
         border-bottom-left-radius: 0px;
         z-index: 101;
     }
-    .count-later-nav {
+    .count-later-nav{
         display: none;
     }
     .input-title {
@@ -671,7 +666,7 @@ a:hover {
     .display-grid-search {
         padding: 2rem 1.8rem;
     }
-    .display-grid-watch-later {
+    .display-grid-watch-later{
         display: grid;
         grid-template-columns: repeat(auto-fit, minmax(270px, 330px));
         gap: 3rem 2rem;
@@ -693,7 +688,7 @@ a:hover {
         transition: button 0.3s ease-out;
         overflow-y: scroll;
         border-radius: 10px;
-        z-index: 101;
+        z-index: 99;
         background-color: transparent;
         padding: 0px;
         background: rgba(255, 255, 255, 0.12);
@@ -742,8 +737,8 @@ a:hover {
         right: unset;
         left: 1%;
         top: 3%;
+        z-index: 99;
         border-radius: 50%;
-        z-index: 102;
         font-size: 20px;
         padding: 10px 10px;
     }
@@ -894,13 +889,13 @@ a:hover {
         -webkit-backdrop-filter: blur(4.7px);
         border: 1px solid rgba(255, 255, 255, 0.27);
     }
-    .cast-actor-grid .li li p {
+    .cast-actor-grid .li li  p {
         margin-top: 1rem;
         font-weight: 800;
         align-self: flex-start;
         color: white;
     }
-    .cast-actor-grid .li li p span {
+    .cast-actor-grid .li li  p span {
         color: #0092ff;
     }
     .genres {
@@ -935,7 +930,7 @@ a:hover {
     .d-flex-header {
         justify-content: space-between;
     }
-    .display-grid-watch-later {
+    .display-grid-watch-later{
         display: grid;
         grid-template-columns: repeat(auto-fit minmax(270px, 330px));
         gap: 3rem 2rem;
@@ -965,7 +960,7 @@ a:hover {
         height: 100%;
         transition: button 0.3s ease-out;
         border-radius: 10px;
-        z-index: 101;
+        z-index: 99;
         background-color: transparent;
         padding: 0px;
         background: rgba(255, 255, 255, 0.12);
@@ -1010,8 +1005,8 @@ a:hover {
         right: unset;
         left: 1%;
         top: 3%;
+        z-index: 99;
         border-radius: 50%;
-        z-index: 102;
         font-size: 20px;
         padding: 10px 10px;
     }
@@ -1114,19 +1109,19 @@ a:hover {
         border-radius: 20px;
         width: 240px;
     }
-    .cast-actor-grid .li li {
+    .cast-actor-grid .li li{
         display: flex;
         flex-direction: column;
         align-items: center;
         padding: 2rem;
         border-radius: 20px;
     }
-    .cast-actor-grid .li li p {
+    .cast-actor-grid .li li  p {
         margin-top: 1rem;
         font-weight: 800;
         align-self: flex-start;
     }
-    .cast-actor-grid .li li p span {
+    .cast-actor-grid .li li  p span {
         color: #0092ff;
     }
     .whole-slider {
@@ -1164,7 +1159,7 @@ a:hover {
         height: 100%;
         transition: button 0.3s ease-out;
         border-radius: 10px;
-        z-index: 101;
+        z-index: 99;
         background-color: transparent;
         padding: 0px;
         background: rgba(255, 255, 255, 0.12);
@@ -1207,8 +1202,8 @@ a:hover {
         right: unset;
         left: 1%;
         top: 3%;
+        z-index: 99;
         border-radius: 50%;
-        z-index: 102;
         font-size: 20px;
         padding: 10px 10px;
     }
@@ -1312,7 +1307,7 @@ a:hover {
         border-radius: 20px;
         width: 240px;
     }
-    .cast-actor-grid .li li {
+    .cast-actor-grid .li li{
         display: flex;
         flex-direction: column;
         align-items: center;
